infra/requirements.txt
aws-cdk-lib==2.182.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==3.8.0
aws-xray-sdk==2.14.0
cdk-nag==2.35.41
boto3==1.37.10
botocore==1.37.10
certifi==2025.1.31
cffi==1.17.1
psycopg[binary]==3.2.5
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.3
responses==0.25.6
urllib3==2.3.0
aws-psycopg2==1.3.8
numpy==2.2.3
pandas==2.2.3